uservoice-to-github/requirements.txt
uservoice>=0.0.23
PyGithub>=1.37